Support `int` index in `MultiEmbeddingTensor` (#198)
